Capstone/w8/main

+++ Controllers/MealPlansController.cs
@@ -2,26 +2,37 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using MyBulkMealsApp.Models;
+using MyBulkMealsApp.Repositories;
+using Newtonsoft.Json;
 
 namespace MyBulkMealsApp.Controllers
 {
+    [Authorize]
     public class MealPlansController : Controller
     {
         private readonly MyBulkMealsAppContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
+        private readonly RecipeRepository _recipes;
 
-        public MealPlansController(MyBulkMealsAppContext context)
+        public MealPlansController(MyBulkMealsAppContext context, UserManager<ApplicationUser> userManager, RecipeRepository recipes)
         {
             _context = context;
+            _userManager = userManager;
+            _recipes = recipes;
         }
+
 
         // GET: MealPlans
         public async Task<IActionResult> Index()
         {
-            return View(await _context.MealPlan.ToListAsync());
+            var userId = _userManager.GetUserId(User);
+            return View(await _context.MealPlan.Where(m => m.CreatorId == userId).ToListAsync());
         }
 
         // GET: MealPlans/Details/5
@@ -33,8 +44,13 @@ namespace MyBulkMealsApp.Controllers
             }
 
             var mealPlan = await _context.MealPlan
+                .Include(m => m.MealPlanEntries)
+                .ThenInclude(r => r.Recipe)
+                .ThenInclude(r => r.Ingredients)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (mealPlan == null)
+            var userId = _userManager.GetUserId(User);
+
+            if (mealPlan == null || mealPlan.CreatorId != userId )
             {
                 return NotFound();
             }
@@ -53,14 +69,33 @@ namespace MyBulkMealsApp.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,PlanName,CreatorId,MealsPerDay,TotalDays,StartDay,EndDay")] MealPlan mealPlan)
+        public async Task<IActionResult> Create([Bind("PlanName,MealsPerDay,TotalDays,StartDay,EndDay")] MealPlan mealPlan, string recipesList)
         {
+            var typeDefinition = new[] { new { Id = 0, Quantity = 0 } };
+            var recipes = JsonConvert.DeserializeAnonymousType(recipesList, typeDefinition);
+
+            List<MealPlanEntry> entries = new List<MealPlanEntry>();
+
+            mealPlan.CreatorId = _userManager.GetUserId(User);
+
+            foreach (var r in recipes)
+            {
+                entries.Add(new MealPlanEntry()
+                {
+                    RecipeId = r.Id,
+                    Quantity = r.Quantity
+                });
+            }
+
             if (ModelState.IsValid)
             {
-                _context.Add(mealPlan);
+                mealPlan.MealPlanEntries = entries;
+                _context.MealPlan.Add(mealPlan);
                 await _context.SaveChangesAsync();
+
                 return RedirectToAction(nameof(Index));
             }
+
             return View(mealPlan);
         }
 
@@ -72,11 +107,33 @@ namespace MyBulkMealsApp.Controllers
                 return NotFound();
             }
 
-            var mealPlan = await _context.MealPlan.FindAsync(id);
-            if (mealPlan == null)
+            var mealPlan = await _context.MealPlan.Where(m => m.Id == id)
+                .Include(m => m.MealPlanEntries)
+                .ThenInclude(m => m.Recipe)
+                .ThenInclude(r => r.Ingredients)
+                .ThenInclude(ri => ri.Ingredient)
+                .FirstOrDefaultAsync();
+
+            var userId = _userManager.GetUserId(User);
+
+            if (mealPlan == null || mealPlan.CreatorId != userId)
             {
                 return NotFound();
             }
+
+            ViewData["RecipesList"] = JsonConvert.SerializeObject(
+                mealPlan.MealPlanEntries.Select(i => new { i.Recipe, i.Quantity })
+                .Select(i => new {
+                    label = i.Recipe.ItemName, //assigning as label lets us use default jquery-ui
+                    i.Recipe.Id,
+                    Calories = i.Recipe.TotalCalories,
+                    Protein = i.Recipe.TotalProtein,
+                    Carbs = i.Recipe.TotalCarbs,
+                    Fat = i.Recipe.TotalFat,
+                    i.Quantity
+                }).ToList(),
+                Formatting.Indented);
+
             return View(mealPlan);
         }
 
@@ -123,9 +180,15 @@ namespace MyBulkMealsApp.Controllers
                 return NotFound();
             }
 
-            var mealPlan = await _context.MealPlan
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (mealPlan == null)
+            var mealPlan = await _context.MealPlan.Where(m => m.Id == id)
+                .Include(m => m.MealPlanEntries)
+                .ThenInclude(r => r.Recipe)
+                .ThenInclude(r => r.Ingredients)
+                .FirstOrDefaultAsync();
+
+            var userId = _userManager.GetUserId(User);
+
+            if (mealPlan == null || mealPlan.CreatorId != userId)
             {
                 return NotFound();
             }
@@ -139,8 +202,14 @@ namespace MyBulkMealsApp.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var mealPlan = await _context.MealPlan.FindAsync(id);
-            _context.MealPlan.Remove(mealPlan);
-            await _context.SaveChangesAsync();
+            var userId = _userManager.GetUserId(User);
+
+            if (mealPlan == null || mealPlan.CreatorId != userId)
+            {
+                _context.MealPlan.Remove(mealPlan);
+                await _context.SaveChangesAsync();
+            }
+
             return RedirectToAction(nameof(Index));
         }
 
@@ -148,5 +217,24 @@ namespace MyBulkMealsApp.Controllers
         {
             return _context.MealPlan.Any(e => e.Id == id);
         }
+
+        [Route("MealPlans/AutocompleteRecipes")]
+        [Route("MealPlans/Edit/AutocompleteRecipes")]
+        public async Task<JsonResult> AutocompleteRecipes(string term)
+        {
+            var foundRecipes = await _recipes.GetByKeywordWithIngredients(term, 6);
+
+            return Json(foundRecipes
+                .Select(i => new {
+                label = i.ItemName, //assigning as label lets us use default jquery-ui
+                i.Id,
+                Calories = i.TotalCalories,
+                Protein = i.TotalProtein,
+                Carbs = i.TotalCarbs,
+                Fat = i.TotalFat
+            }).ToArray());
+        }
+
+
     }
 }
